perplexity/requirements.txt
pillow>=10.0
numpy>=1.26
scikit-image>=0.22
faiss-cpu>=1.8
tqdm>=4.66
ipywidgets>=8.1
matplotlib>=3.8
pytest>=8.0
jupyter>=1.0
notebook>=7.0
